added image transpose feature
